requirements.txt
Flask
Pillow
captcha
requests
flask-sqlalchemy